Request for search fields sequentially
This is an hypothetical fix as the bug is hard to reproduce,
so it's difficult to know if the fix is indeed having an effect.
On some installations of Katsu, namely in development mode, using
Django's built-in server, and possibly a debugger, the call for search fields
is throttled and during that time, no other request can be made on Katsu.
This call also eventually fails due to Gateway timeout.
It is possible that this throttling is induced by the quick succession
of API calls, on a monothreaded instance (?).
The rationale behind the fix is to avoid concurrency issues by performing
the two API calls aimed at Katsu sequentially.